js/css using relative paths and merging demo with sample app
