+++ src/samples/header.php
@@ -1,0 +1,69 @@
+<?php
+/**
+ * Copyright (c) 2016 Micorosft Corporation
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ *
+ * @author Aashay Zajriya <aashay@introp.net>
+ * @license MIT
+ * @copyright (C) 2016 onwards Microsoft Corporation (http://microsoft.com/)
+ */
+?>
+<head>
+    <?php if (isset($_SERVER['HTTP_HOST'])) { 
+        $hostname = $_SERVER['HTTP_HOST'];
+     } else {
+         $hostname = '';
+     }?>
+    <link rel="stylesheet" type="text/css" href="<?php $hostname ?>/css/bootstrap.css"/>
+    <link rel="stylesheet" type="text/css" href="<?php $hostname ?>/css/bootstrap-theme.css"/>
+    <script type="text/javascript" src="<?php $hostname ?>/js/jquery-1.10.2.js"></script>
+    <script type="text/javascript" src="<?php $hostname ?>/js/jquery.validate.js"></script>
+    <script type="text/javascript" src="<?php $hostname ?>/js/bootstrap.js"></script>
+    <style type="text/css">
+        body {
+            padding-top: 50px;
+        }
+        .starter-template {
+            padding: 40px 15px;
+            text-align: center;
+        }
+    </style>
+</head>
+<body>
+    <nav class="navbar navbar-inverse navbar-fixed-top">
+        <div class="container">
+            <div class="navbar-header">
+                <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#navbar" aria-expanded="false" aria-controls="navbar">
+                    <span class="sr-only">Toggle navigation</span>
+                    <span class="icon-bar"></span>
+                    <span class="icon-bar"></span>
+                    <span class="icon-bar"></span>
+                </button>
+                <a class="navbar-brand" href="/">ADAL PHP Demo</a>
+            </div>
+            <?php if (isset($_SESSION['user_id'])) { ?>
+                <div id="navbar" class="collapse navbar-collapse">
+                    <ul class="nav navbar-nav pull-right">
+                        <li><a href="/logout.php">Logout</a></li>
+                    </ul>
+                </div><!--/.nav-collapse -->
+            <?php } ?>
+        </div>
+    </nav>
