+++ src/samples/js/jquery-1.10.2.js
